🚚 move chromatic stories to the __tests__ folders
