+++ src/components/alert/__tests__/alert.stories.spec.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Alert from './Alert';
+import Alert from '../Alert';
 
 export default {
   component: Alert,
