+++ src/components/bullet/__tests__/bullet.stories.spec.tsx
@@ -1,7 +1,7 @@
 import React, { ReactNode } from 'react';
 import { ComponentMeta, ComponentStory } from '@storybook/react';
 
-import Bullet from './Bullet';
+import Bullet from '../Bullet';
 
 export default {
   component: Bullet,
