+++ src/components/datagrid/__tests__/datagrid.stories.spec.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 
-import DataGrid from './DataGrid';
-import Link from '../link';
-import { IconMenu, MenuItem } from '../menu';
-import { rows1 } from '../../static/data/datagrid';
+import DataGrid from '../DataGrid';
+import Link from '../../link';
+import { IconMenu, MenuItem } from '../../menu';
+import { rows1 } from '../../../static/data/datagrid';
 import { ComponentMeta, ComponentStory } from '@storybook/react';
 
 export default {
